requirements.txt
annotated-types==0.7.0
easydict==1.13
pip==23.2.1
pydantic==2.7.4
pydantic_core==2.18.4
pyroute2==0.7.12
PyYAML==6.0.1
typing_extensions==4.12.2
